Move Lens functional tests to x-pack
