+++ test/functional/config.js
@@ -35,7 +35,6 @@ export default async function ({ readConfigFile }) {
       require.resolve('./apps/status_page'),
       require.resolve('./apps/timelion'),
       require.resolve('./apps/visualize'),
-      require.resolve('./apps/lens'),
     ],
     pageObjects,
     services,
@@ -77,9 +76,6 @@ export default async function ({ readConfigFile }) {
       visualize: {
         pathname: '/app/kibana',
         hash: '/visualize',
-      },
-      lens: {
-        pathname: '/app/lens',
       },
       dashboard: {
         pathname: '/app/kibana',
