+++ x-pack/test/functional/config.js
@@ -34,6 +34,7 @@ export default async function ({ readConfigFile }) {
       resolve(__dirname, './apps/discover'),
       resolve(__dirname, './apps/security'),
       resolve(__dirname, './apps/spaces'),
+      resolve(__dirname, './apps/lens'),
       resolve(__dirname, './apps/logstash'),
       resolve(__dirname, './apps/grok_debugger'),
       resolve(__dirname, './apps/infra'),
@@ -99,6 +100,9 @@ export default async function ({ readConfigFile }) {
     // Kibana's config in order to use this helper
     apps: {
       ...kibanaFunctionalConfig.get('apps'),
+      lens: {
+        pathname: '/app/lens',
+      },
       login: {
         pathname: '/login',
       },
